overhauled game entities to use floats for their physical properties

+++ Engine/Dude.h
@@ -9,14 +9,14 @@ public:
 	void ClampToScreen();
 	void Draw( Graphics& gfx ) const;
 	void Update( const Keyboard& kbd );
-	int GetX() const;
-	int GetY() const;
-	int GetWidth() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetWidth() const;
+	float GetHeight() const;
 private:
-	int x = 400;
-	int y = 300;
-	static constexpr int speed = 1;
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	float x = 400.0f;
+	float y = 300.0f;
+	static constexpr float speed = 1.0f;
+	static constexpr float width = 20.0f;
+	static constexpr float height = 20.0f;
 };
